Made models multiple output
