+++ glupredkit/helpers/tf_keras.py
@@ -3,27 +3,55 @@ import numpy as np
 from glupredkit.helpers.model_config_manager import ModelConfigurationManager
 
 
-def prepare_sequences(data, labels, window_size, real_time, step_size=12):
+def prepare_sequences(df_X, df_y, window_size, what_if_columns, prediction_horizon, real_time, step_size=2):
     X, y, dates = [], [], []
-    exclude_list = ["target", "imputed"]
-    sequence_columns = [item for item in data.columns if item not in exclude_list]
+    target_columns = df_y.columns
+    exclude_list = list(target_columns) + ["imputed", "iob", "cob", "carbs"]
+    sequence_columns = [item for item in df_X.columns if item not in exclude_list]
+    n_what_if = prediction_horizon // 5
 
-    for i in range(0, len(data) - window_size, step_size):
-        sequence = data[sequence_columns][i:i + window_size]
-        label = labels.iloc[i + window_size - 1]
-        date = labels.index[i + window_size - 1]
+    print("Preparing sequences...")
 
-        if 'imputed' in data.columns:
-            imputed = data['imputed'].iloc[i + window_size - 1]
+    for i in range(0, len(df_X) - window_size - n_what_if, step_size):
+        label = df_y.iloc[i + window_size - 1]
+
+        if df_X.iloc[i:i + window_size + n_what_if].isnull().any().any():
+            continue  # Skip this sequence if there are NaN values in the input data
+
+        if (df_X[['cob']].iloc[i + window_size:i + window_size + n_what_if] > 0).any().any():
+            continue  # Skip this sequence if there are NaN values in the input data
+
+        if 'imputed' in df_X.columns:
+            imputed = df_X['imputed'].iloc[i + window_size - 1]
             if imputed:
                 continue  # Skip this sequence
 
         if not real_time:
-            if pd.isna(label):
+            if pd.isna(label).any():
                 continue  # Skip this sequence
 
-        X.append(sequence)
-        y.append(label)
+        # sequence = df_X[sequence_columns][i:i + window_size]
+        date = df_y.index[i + window_size - 1]
+
+        # Initialize an empty DataFrame for mixed_sequence with the same columns as sequence
+        mixed_sequence = pd.DataFrame(index=df_X[i:i + window_size + n_what_if].index, columns=sequence_columns).iloc[0:window_size + n_what_if]
+
+        for col in sequence_columns:
+            if col in what_if_columns:
+                # For "what if" columns, extend the sequence
+                mixed_sequence[col] = df_X[col][i:i + window_size + n_what_if]
+            else:
+                # Slice the original window size for other columns
+                original_sequence = df_X[col][i:i + window_size]
+                # Create a padding series with -1 values and the appropriate date index
+                padding_index = df_X.index[i + window_size:i + window_size + n_what_if]
+                padding_series = pd.Series([-1] * n_what_if, index=padding_index)
+                # Concatenate the original sequence with the padding series
+                full_sequence = pd.concat([original_sequence, padding_series])
+                mixed_sequence[col] = full_sequence
+
+        X.append(mixed_sequence)
+        y.append(label.values.tolist())
         dates.append(date)
 
     return np.array(X), np.array(y), dates
